Wode's Grimoire: fix Grace
Simplified Grace's effect function.
The addModifier gets called for both prediction and regular so it
doesn't need to be wrapped in an if function checking for prediction.
The sfx however does, and the sfx should only be called once since
it should play only once no matter how many targets there are.

modifierGraceCountdown and castingGracecounter (renamed for clarity to
graceCountdown and graceQuantity respectively) need to be a part of
the modifiers state, otherwise they are global to the whole module.