Wode's Grimoire: Fix FlameStrike
getUnitsWithinDistanceOfTarget will also return the target
if the target is a unit (target can be any Vec2).
So I had to refactor the damage logic.  Previously it
was dealing too much to the `unit` (40+10) and if
you used target circle + flamestrike and the circle
included the player it wouldn't deal damage to the player
at all (I think because of the .splice)